requirements.txt
boto3>=1.26.0
pandas>=1.5.0
numpy>=1.24.0
sagemaker>=2.160.0
tqdm>=4.64.0
python-dotenv>=1.0.0
